requirements.txt
streamlit
paddleocr
pytesseract
python-pptx
pdf2image
opencv-python-headless
numpy
wordfreq
rapidfuzz
pillow
paddlepaddle
extract-msg
tkinterdnd2